+++ test/all_test.dart
@@ -1,9 +1,11 @@
+import 'inflection_test.dart' as inflection_test;
 import 'plural_test.dart' as plural_test;
 import 'singular_test.dart' as singular_test;
 import 'snake_case_test.dart' as snake_case_test;
 import 'spinal_case_test.dart' as spinal_case_test;
 
 void main() {
+  inflection_test.main();
   plural_test.main();
   singular_test.main();
   snake_case_test.main();